Split ConsumerStage file into more dedicated source files
